Enable new processmanager, add Tilt support
