+++ circus.ini
@@ -13,7 +13,7 @@ numprocesses = 2
 copy_env = True
 virtualenv = $(circus.env.HOME)/venv
 working_dir = $(circus.env.HOME)/fermentrack
-# Log ouput, rotate every 2MB, keep 5 logfiles.
+# Log output, rotate every 2MB, keep 5 logfiles.
 stderr_stream.class = FileStream
 stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack-stderr.log
 stderr_stream.time_format = %Y-%m-%d %H:%M:%S
@@ -37,39 +37,21 @@ port = 8123
 PYTHONPATH = /home/fermentrack/fermentrack
 
 
+[watcher:processmgr]
+cmd = $(circus.env.HOME)/venv/bin/python
+args = $(circus.env.HOME)/fermentrack/utils/processmgr.py
+virtualenv = $(circus.env.HOME)/venv
+working_dir = $(circus.env.HOME)/fermentrack
+copy_env = True
 
-#[watcher:brewpi-spawner]
-#cmd = $(circus.env.HOME)/venv/bin/python
-#args = $(circus.env.HOME)/fermentrack/utils/brewpi_spawner.py
-#virtualenv = $(circus.env.HOME)/venv
-#working_dir = $(circus.env.HOME)/fermentrack
-#copy_env = True
-
-#use_sockets = False
-#numprocesses = 1
-# Log ouput, rotate every 2MB, keep 5 logfiles.
-#stderr_stream.class = FileStream
-#stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack-brewpi-spawner.log
-#stderr_stream.time_format = %Y-%m-%d %H:%M:%S
-#stderr_stream.max_bytes = 2097152
-#stderr_stream.backup_count = 5
-
-
-#[watcher:processmgr_new]
-#cmd = $(circus.env.HOME)/venv/bin/python
-#args = $(circus.env.HOME)/fermentrack/utils/processmgr_new.py
-#virtualenv = $(circus.env.HOME)/venv
-#working_dir = $(circus.env.HOME)/fermentrack
-#copy_env = True
-
-#use_sockets = False
-#numprocesses = 1
-# Log ouput, rotate every 2MB, keep 5 logfiles.
-#stderr_stream.class = FileStream
-#stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack-processmgr.log
-#stderr_stream.time_format = %Y-%m-%d %H:%M:%S
-#stderr_stream.max_bytes = 2097152
-#stderr_stream.backup_count = 5
+use_sockets = False
+numprocesses = 1
+# Log output, rotate every 2MB, keep 5 logfiles.
+stderr_stream.class = FileStream
+stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/fermentrack-processmgr.log
+stderr_stream.time_format = %Y-%m-%d %H:%M:%S
+stderr_stream.max_bytes = 2097152
+stderr_stream.backup_count = 5
 
 [watcher:huey]
 cmd = $(circus.env.HOME)/venv/bin/python
@@ -81,7 +63,7 @@ copy_env = True
 
 use_sockets = False
 numprocesses = 1
-# Log ouput, rotate every 2MB, keep 5 logfiles.
+# Log output, rotate every 2MB, keep 5 logfiles.
 stderr_stream.class = FileStream
 stderr_stream.filename = $(circus.env.HOME)/fermentrack/log/huey-stderr.log
 stderr_stream.refresh_time = 0.3
@@ -93,5 +75,3 @@ stdout_stream.filename = $(circus.env.HOME)/fermentrack/log/huey-stdout.log
 stdout_stream.refresh_time = 0.3
 stdout_stream.max_bytes = 2097152
 stdout_stream.backup_count = 5
-stderr_stream.max_bytes = 2097152
-stderr_stream.backup_count = 5
